pyproject.toml
[project]
name = "pomoco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bokeh>=3.8.1",
    "brainsets",
    "dandi>=0.73.2",
    "holoviews>=1.22.0",
    "ipykernel>=7.1.0",
    "jax>=0.8.1",
    "matplotlib>=3.10.7",
    "pytorch-brain",
    "ratinabox>=1.15.3",
]

[tool.uv.workspace]
members = [
    "brainsets",
    "torch_brain",
]

[tool.uv.sources]
brainsets = { workspace = true, editable = true }
pytorch-brain = { workspace = true, editable = true }
